Create GlobalEsp_and_aimbot
Meant to be open source and a contribution to the Roblox developers community

J   - make world brighter
`   - Aim-bot, locks onto the nearest player to your mouse

When you die the esp breaks idk what to do about it tho but the aimbot will work fine 

This script was tested on Project delta and by all means meant to be used for educational purposes but what you do with it does not matter me and I am not liable for any bans, loss of Ingame items or whatever happens. You are responsible for using it and not me.